Move TestAdminCommandsFailFast to separate file
test_admin.py requires OMERODIR to be set for all tests due to an autouse fixture https://github.com/ome/omero-py/blob/b1a627eba9a7dce7880645fd0a42d85949851ff7/test/unit/clitest/test_admin.py#L51-L52
